+++ xmake/actions/config/configheader.lua
@@ -133,7 +133,7 @@ function make()
     local targetname = option.get("target")
 
     -- enter project directory
-    local olddir = os.cd(project.directory())
+    local oldir = os.cd(project.directory())
 
     -- init files
     local files = {}
@@ -154,5 +154,5 @@ function make()
     end
  
     -- leave project directory
-    os.cd(olddir)
+    os.cd(oldir)
 end