+++ xmake/actions/package/main.lua
@@ -142,7 +142,7 @@ end
 function _package(target)
 
     -- enter project directory
-    local olddir = os.cd(project.directory())
+    local oldir = os.cd(project.directory())
 
     -- the target scripts
     local scripts =
@@ -161,7 +161,7 @@ function _package(target)
     end
 
     -- leave project directory
-    os.cd(olddir)
+    os.cd(oldir)
 end
 
 -- package the given target and deps