+++ xmake/plugins/project/main.lua
@@ -36,6 +36,7 @@ import("vstudio.vs2012")
 import("vstudio.vs2013")
 import("vstudio.vs2015")
 import("vstudio.vs2017")
+import("clang.compile_commands")
 
 -- make project
 function _make(kind)
@@ -43,16 +44,17 @@ function _make(kind)
     -- the maps
     local maps = 
     {
-        makefile    = makefile.make
-    ,   vs2002      = vs2002.make
-    ,   vs2003      = vs2003.make
-    ,   vs2005      = vs2005.make
-    ,   vs2008      = vs2008.make
-    ,   vs2010      = vs2010.make
-    ,   vs2012      = vs2012.make
-    ,   vs2013      = vs2013.make
-    ,   vs2015      = vs2015.make
-    ,   vs2017      = vs2017.make
+        makefile         = makefile.make
+    ,   vs2002           = vs2002.make
+    ,   vs2003           = vs2003.make
+    ,   vs2005           = vs2005.make
+    ,   vs2008           = vs2008.make
+    ,   vs2010           = vs2010.make
+    ,   vs2012           = vs2012.make
+    ,   vs2013           = vs2013.make
+    ,   vs2015           = vs2015.make
+    ,   vs2017           = vs2017.make
+    ,   compile_commands = compile_commands.make
     }
     assert(maps[kind], "the project kind(%s) is not supported!", kind)